Fix location coordinates to read single bytes at correct offsets
Per madmoose dune-rust data.rs Sietch structure:
- map_x is at offset 3 (single byte)
- map_y is at offset 4 (single byte)

Changed GetLocationCoordinates to return (byte, byte) instead of (ushort, ushort)
and read from offsets 3 and 4 instead of 2 and 4.

+++ src/Cryogenic/GameEngineWindow/Models/DuneGameState.cs
@@ -237,7 +237,7 @@ public partial class DuneGameState : MemoryBasedDataStructure {
 
     public byte GetSietchStatus(int index) => GetLocationStatus(index);
     public ushort GetSietchSpiceField(int index) => (ushort)GetLocationSpiceAmount(index);
-    public (ushort X, ushort Y) GetSietchCoordinates(int index) => GetLocationCoordinates(index);
+    public (byte X, byte Y) GetSietchCoordinates(int index) => GetLocationCoordinates(index);
     
     /// <summary>
     /// Returns the count of discovered locations (including palaces, villages, sietches).
